# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RaahSetu routing API",
    "description": "SIH prototype. Custom risk-aware A*. No live navigation guarantees.",
    "version": "0.1.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/api/v1/bootstrap": {
      "get": {
        "summary": "Bootstrap",
        "operationId": "bootstrap_api_v1_bootstrap_get",
        "parameters": [
          {
            "name": "dataset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo",
              "title": "Dataset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/terrain": {
      "get": {
        "summary": "Terrain",
        "operationId": "terrain_api_v1_terrain_get",
        "parameters": [
          {
            "name": "dataset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo",
              "title": "Dataset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/network": {
      "get": {
        "summary": "Network",
        "operationId": "network_api_v1_network_get",
        "parameters": [
          {
            "name": "weather",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "normal",
              "title": "Weather"
            }
          },
          {
            "name": "dataset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "demo",
              "title": "Dataset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/routes/compare": {
      "post": {
        "summary": "Compare Routes",
        "operationId": "compare_routes_api_v1_routes_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Endpoint": {
        "properties": {
          "node_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Node Id"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lon": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Lon"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "Endpoint"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "RouteRequest": {
        "properties": {
          "dataset_id": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Dataset Id",
            "default": "demo"
          },
          "origin": {
            "$ref": "#/components/schemas/Endpoint"
          },
          "destination": {
            "$ref": "#/components/schemas/Endpoint"
          },
          "vehicle": {
            "type": "string",
            "enum": [
              "heavy",
              "emergency",
              "light"
            ],
            "title": "Vehicle",
            "default": "heavy"
          },
          "risk_aversion": {
            "type": "number",
            "maximum": 5.0,
            "minimum": 0.0,
            "title": "Risk Aversion",
            "default": 1.5
          },
          "weather": {
            "type": "string",
            "enum": [
              "normal",
              "heavy_rain"
            ],
            "title": "Weather",
            "default": "normal"
          },
          "closed_edge_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 500,
            "title": "Closed Edge Ids"
          },
          "strict_vehicle": {
            "type": "boolean",
            "title": "Strict Vehicle",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "origin",
          "destination"
        ],
        "title": "RouteRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}